Load more links when reaching bottom of recycler list
Implemented an onscroll listener that checks to see if there are about 5
or less links to scroll to in the recyclerview.  One drawback of the
current implementation is that it will call the getLinks function
several times.  I've ensured that the api will only get pinged once by
implementing a boolean gate on the getLinks() function, I'm not sure if
there's a cleaner way to get implement the onscroll listener.